correct the name of landing & accoutn page
